+++ packages/istanbul-widget/src/core/plugin/IstanbulWidgetReactPlugin.ts
@@ -1,15 +1,14 @@
 import React from 'react'
-import ReactDOM from 'react-dom/client'
 import { $ } from '@/utils/query'
 import { IstanbulWidget } from '../core'
+import { reactdomRender, reactdomUnmount } from '../dom/react-render'
 import { type PluginType } from '../options.interface'
 import { IstanbulWidgetPlugin } from './IstanbulWidgetPlugin'
 
 export type IstanbulWidgetReactPluginProps = {} & PluginType
 
 export class IstanbulWidgetReactPlugin<T extends {} = {}> extends IstanbulWidgetPlugin {
-  private _root!: ReactDOM.Root
-
+  private _root!: HTMLDivElement
   constructor(
     /**
      * 插件id
@@ -35,16 +34,18 @@ export class IstanbulWidgetReactPlugin<T extends {} = {}> extends IstanbulWidget
   onRender() {
     this.on('render', () => {
       const el = document.createElement('div')
-      const root = ReactDOM.createRoot(el)
-      root.render(
+
+      this._root = el
+
+      reactdomRender(
         React.createElement(this.Component, {
           ...((this.initialProps || {}) as T),
           id: this.id,
           name: this.name,
           domID: this.domID,
         }),
+        el,
       )
-      this._root = root
 
       const target = $.queryEl(`#${this.domID}`)
       target.appendChild(el)
@@ -57,7 +58,7 @@ export class IstanbulWidgetReactPlugin<T extends {} = {}> extends IstanbulWidget
     if (!this._root) {
       IstanbulWidget.logger.warn('[istanbul-widget]: init component first')
     } else {
-      this._root.unmount()
+      reactdomUnmount(this._root)
     }
   }
 }